+++ CHANGELOG.md
@@ -5,8 +5,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 
-## [1.2.0] - 2025-05-31
+## [1.3.0] - 2025-05-31
 - Adding a prompt to choose the language type between typescript and javascript. 
+- Adding a 3rd choice into the architecture prompt.
+- Create folders structure depends the architecture type.
 
 ## [1.1.0] - 2025-05-30
 - Adding a first prompt to the project to enter the name of the project with command `./archi create`